Split schema file
