+++ packages/db/src/schema/tags.ts
@@ -1,0 +1,18 @@
+import { relations } from "drizzle-orm";
+import { jsonb, pgTable, text, timestamp } from "drizzle-orm/pg-core";
+
+import { projectsToTags } from "./projects";
+
+export const tags = pgTable("tags", {
+  id: text("id").primaryKey(),
+  code: text("code").notNull().unique(),
+  name: text("name").notNull(),
+  description: text("description"),
+  aliases: jsonb("aliases"),
+  createdAt: timestamp("created_at").notNull().defaultNow(),
+  updatedAt: timestamp("updated_at"),
+});
+
+export const tagsRelations = relations(tags, ({ many }) => ({
+  projectsToTags: many(projectsToTags),
+}));
